refactor/remove_display_manager
display manager is an early implementation aimed at the mark1
it has since been replaced more than once:
 - active skills in intent stage
 - idle screen manager for the mk2 (skill)

DisplayManager is virtually unused, afaik nothing uses it and even if something does parse ipc or import this class it is not getting accurate data

not all methods of enclosure api make a skill active

this implementation was abandoned before it was ever finished

DisplayManager will be refactored and ported into PHAL
active skills (in the sense of interacting with hardware) will be tracked there
