+++ mycroft/enclosure/display_manager.py
@@ -13,6 +13,20 @@
 # limitations under the License.
 #
 
+
+"""
+NOTE: this is dead code! do not use!
+
+This file is only present to ensure backwards compatibility
+in case someone is importing from here
+
+This is only meant for 3rd party code expecting ovos-core
+to be a drop in replacement for mycroft-core
+
+TODO: consider importing from PHAL
+"""
+
+
 """ DisplayManager
 
 This module provides basic "state" for the visual representation associated
